Refactor test data
